==> PD1_Treat/01a_SS2/Immune/02_ProjectBroad.DetailedAnnots.Rmd <==
---
title: "ProjectBroadAnnots.Rmd"
author: "Jenna LaBelle"
date: "10/07/22"
output: html_document
---

Project "broad" (T cell and myeloid) and "detailed" (CD4, CD8, DC, Monocyte, Bcell) annotations from the full immune cohort to PD1 treated patients

```{r}
library(Seurat) 

base_dir = "/Users/jlabelle/"
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Shared/Scripts/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/PD1_Treat/01a_SS2/Immune/")

analysis_dir<- paste0(working_dir, "/analysis/project_broad.detailed_annot/")
figure_dir<- paste0(working_dir, "/figures/project_broad.detailed_annot/")
full_cohort_dir<- "../../../ImmuneCellAnalysis/"

if(!dir.exists(figure_dir)){dir.create(figure_dir, recursive = TRUE)}
if(!dir.exists(analysis_dir)){dir.create(analysis_dir, recursive = TRUE)}

preprocessing = paste0(base_dir, script_dir, "single_cell_preprocessing_helper_functions.R")
nmf_helper = paste0(base_dir, script_dir, "NMF_helper_function.R")
source(preprocessing)
source(nmf_helper)
source(paste0(base_dir, script_dir, "seurat_transferAnchors_helper.R"))
```

## Load seurat objects: pd-1 (processed in 01b_Preprocess_QC_ToSeurat.Rmd) and full immune cohort (both myeloid and t cells)
```{r}
## load seurat objects
pd1_seurat<- readRDS("analysis/preprocess/seurat_obj.Rds")
full_seurat<- readRDS(paste0(full_cohort_dir,
                             "01_Preprocessing/02_Counts_Preprocess/analysis/detailed_annotation/pedOnly_nomuv63/",
                             "tcell.myeloid_seurat_harmony.Rds"))
full_tcell_seurat<- readRDS(paste0(full_cohort_dir,
                                   "02a_Tcells/02_Ped.Adult/03_Visualize.Reannotate_programs/analysis/Reannotate_programs/",
                                   "rank8_sharedByunion/",
                                   "ped_seurat_tcell_harmony.Rds"))
full_myeloid_seurat<- readRDS(paste0(full_cohort_dir,
                                     "01_Preprocessing/02_Counts_Preprocess/analysis/detailed_annotation/",
                                     "pedOnly_nomuv63/",
                                     "myeloid_seurat_harmony.Rds"))

## Colors 
immune_colors<- readRDS(paste0(full_cohort_dir, "/plot_colors/ImmuneCell.celltypes.Rds"))

```

## Projection of myeloid/tcell annotations with pcproject 
```{r}
pd1_seurat<- my_project_seurat(full_seurat, pd1_seurat, reference_annotation = "broad_annot",
                               prediction_colNames = c("broad_annot_projected", "Tcell_score_projected", 
                                                             "Myeloid_score_projected", "max_score_projected"))


## plot
DimPlot(pd1_seurat, group.by="broad_annot_projected", pt.size = 1, label=TRUE, label.size = 6)+ NoLegend()+
  scale_color_manual(values=immune_colors[names(immune_colors) %in% pd1_seurat$broad_annot_projected])
ggsave(paste0(figure_dir, "UMAP_broadAnnot_projectedFromFull.png"), width=6, height=6)

## Confirm high expression of myeloid/tcell markers by predicted broad annotation
cowplot::plot_grid(plotlist=lapply(c("CSF1R", "CD3G", "CD8A", "CD4"), function(x){
  VlnPlot(pd1_seurat, features=x, group.by = "broad_annot_projected")+ NoLegend()+
    scale_fill_manual(values=immune_colors[names(immune_colors) %in% pd1_seurat$broad_annot_projected])+
    xlab("")
}), ncol=4)
ggsave(paste0(figure_dir, "Vln_broadMarkers_projectedFromFull.png"), width=12, height=4)

## Proportion
my_barchart_programProp(pd1_seurat, coloring_variable = "broad_annot_projected", colors=immune_colors)
ggsave(paste0(figure_dir, "Barchart_broadannot_bySample.png"), width=4, height=4)

saveRDS(pd1_seurat, "analysis/preprocess/seurat_obj.Rds")
```

## Split into myeloid/tcell seurat objects, rerun processing
```{r}
## Myeloid
myeloid_pd1<- subset(pd1_seurat, broad_annot_projected=="Myeloid")
myeloid_pd1_meta<- myeloid_pd1@meta.data
myeloid_pd1<- RunFullSeurat(myeloid_pd1@assays$RNA@counts, samples= myeloid_pd1$sample, RunHarmony = TRUE)
myeloid_pd1<- AddMetaData(myeloid_pd1, myeloid_pd1_meta[,!colnames(myeloid_pd1_meta) %in% colnames(myeloid_pd1@meta.data)])

## Tcell
tcell_pd1<- subset(pd1_seurat, broad_annot_projected=="Tcell")
tcell_pd1_meta<- tcell_pd1@meta.data
tcell_pd1<- RunFullSeurat(tcell_pd1@assays$RNA@counts, samples= tcell_pd1$sample, RunHarmony = TRUE)
tcell_pd1<- AddMetaData(tcell_pd1, tcell_pd1_meta[,!colnames(tcell_pd1_meta) %in% colnames(tcell_pd1@meta.data)])

saveRDS(myeloid_pd1, file=paste0(analysis_dir, "myeloid_seurat.Rds"))
saveRDS(tcell_pd1, file=paste0(analysis_dir, "tcell_seurat.Rds"))
```


## Projection of DC/BC/Monocyte annotations for myeloid cells with pcproject 
```{r}
myeloid_pd1<- readRDS(paste0(analysis_dir, "myeloid_seurat.Rds"))

myeloid_pd1<- my_project_seurat(full_myeloid_seurat, myeloid_pd1, reference_annotation = "detailed_annot",
                                prediction_colNames =c("detailed_annot_projected", "Myeloid_score_projected", 
                                                             "Bcell_score_projected", "DC_score_projected",
                                                             "max_Detailed_score_projected") )


## plot
DimPlot(myeloid_pd1, group.by="detailed_annot_projected", pt.size = 1, label=TRUE, label.size = 6)+ NoLegend()+
  scale_color_manual(values=immune_colors[names(immune_colors) %in% myeloid_pd1$detailed_annot_projected])
ggsave(paste0(figure_dir, "UMAP_myeloid_detailedAnnot_projectedFromFull.png"), width=6, height=6)

  
## Confirm high expression of myeloid markers by predicted detailed annotation
cowplot::plot_grid(plotlist=lapply(c( "CSF1R",  "CD207","CD19","IL3RA", "CLEC4C", "CD3G"), function(x){
  VlnPlot(myeloid_pd1, features=x, group.by = "detailed_annot_projected")+ NoLegend()+
    scale_fill_manual(values=immune_colors[names(immune_colors) %in% myeloid_pd1$detailed_annot_projected])+
    xlab("")
}), ncol=3)
ggsave(paste0(figure_dir, "Vln_myeloid_detailedMarkers_projectedFromFull.png"), width=7, height=8)


## "Bcells" seem to be plastomcytoid DC. Rename
myeloid_pd1$detailed_annot_projected_manualPDC<- gsub("BCell", "pDC", myeloid_pd1$detailed_annot_projected)

## Plot updated annots
DimPlot(myeloid_pd1, group.by="detailed_annot_projected_manualPDC", pt.size = 1, label=TRUE, label.size = 6)+ NoLegend()+
  scale_color_manual(values=c(immune_colors[names(immune_colors) %in% myeloid_pd1$detailed_annot_projected],
                              pDC="pink"))

  ## Umap of annots
  ggsave(paste0(figure_dir, "UMAP_myeloid_detailedAnnot_projectedFromFull_manualpDC.png"), width=6, height=6)

  ## Confirm expression with vln
  cowplot::plot_grid(plotlist=lapply(c( "CSF1R",  "CD207","CD19","IL3RA", "CLEC4C", "CD3G"), function(x){
    VlnPlot(myeloid_pd1, features=x, group.by = "detailed_annot_projected_manualPDC")+ NoLegend()+
      scale_fill_manual(values=c(immune_colors[names(immune_colors) %in% myeloid_pd1$detailed_annot_projected],
                              pDC="pink"))+
      xlab("")
  }), ncol=3)
  ggsave(paste0(figure_dir, "Vln_myeloid_detailedMarkers_projectedFromFull_manualpDC.png"), width=7, height=8)


## Save with annotations
saveRDS(myeloid_pd1, file=paste0(analysis_dir, "myeloid_seurat.Rds"))

## Remove DC/BC, rerun clustering, and save
myeloid_pd1_noDC.B<- subset(myeloid_pd1, detailed_annot_projected_manualPDC=="Myeloid")
myeloid_pd1_noDC.B_meta<- myeloid_pd1_noDC.B@meta.data
myeloid_pd1_noDC.B<- RunFullSeurat(myeloid_pd1_noDC.B@assays$RNA@counts, RunHarmony = FALSE, samples=myeloid_pd1_noDC.B$sample)
myeloid_pd1_noDC.B<- AddMetaData(myeloid_pd1_noDC.B, myeloid_pd1_noDC.B_meta[,!colnames(myeloid_pd1_noDC.B_meta) %in% 
                                                                               colnames(myeloid_pd1_noDC.B@meta.data)])
saveRDS(myeloid_pd1_noDC.B, file=paste0(analysis_dir, "myeloid_noDC.B_seurat.Rds"))
```




## Projection of CD4/CD8 annotations for T cells with pcproject 
```{r}
tcell_pd1<- readRDS(paste0(analysis_dir, "tcell_seurat.Rds"))

tcell_pd1<- my_project_seurat(full_tcell_seurat, tcell_pd1, reference_annotation = "detailed_annot_byNMFrank8_mergedByunion",
                                prediction_colNames =c("detailed_annot_projected", "CD4_score_projected", 
                                                             "CD8_score_projected", "max_Detailed_score_projected"))


## plot
DimPlot(tcell_pd1, group.by="detailed_annot_projected", pt.size = 1, label=TRUE, label.size = 6)+ NoLegend()+
  scale_color_manual(values=immune_colors[names(immune_colors) %in% tcell_pd1$detailed_annot_projected])
ggsave(paste0(figure_dir, "UMAP_Tcell_detailedAnnot_projectedFromFull.png"), width=6, height=6)

## Confirm high expression of tcell markers by predicted detailed annotation
cowplot::plot_grid(plotlist=lapply(c( "CD3G", "CD8A", "CD4"), function(x){
  VlnPlot(tcell_pd1, features=x, group.by = "detailed_annot_projected")+ NoLegend()+
    scale_fill_manual(values=immune_colors[names(immune_colors) %in% tcell_pd1$detailed_annot_projected])+
    xlab("")
}), ncol=3)
ggsave(paste0(figure_dir, "Vln_Tcell_detailedMarkers_projectedFromFull.png"), width=9, height=4)

FeaturePlot(tcell_pd1, features=c("CD3G", "CD4", "CD8A"), cols = c("grey", "red"), ncol=3)
ggsave(paste0(figure_dir, "UMAP_Tcell_detailedMarkers.png"), width=12, height=3)

## Save with annotations
saveRDS(tcell_pd1, file=paste0(analysis_dir, "tcell_seurat.Rds"))

## Split into cd4/cd8, recluster
cd4_pd1<- subset(tcell_pd1, detailed_annot_projected=="CD4")
cd4_pd1_meta<- cd4_pd1@meta.data
cd4_pd1<- RunFullSeurat(cd4_pd1@assays$RNA@counts, samples=cd4_pd1$sample, RunHarmony = FALSE)
cd4_pd1<- AddMetaData(cd4_pd1, cd4_pd1_meta[,!colnames(cd4_pd1_meta) %in% colnames(cd4_pd1@meta.data)])

cd8_pd1<- subset(tcell_pd1, detailed_annot_projected=="CD8")
cd8_pd1_meta<- cd8_pd1@meta.data
cd8_pd1<- RunFullSeurat(cd8_pd1@assays$RNA@counts, samples=cd8_pd1$sample, RunHarmony = FALSE)
cd8_pd1<- AddMetaData(cd8_pd1, cd8_pd1_meta[,!colnames(cd8_pd1_meta) %in% colnames(cd8_pd1@meta.data)])

## Both CD4 and CD8 show misannotated cells- fix annotation and recluster
p1<-FeaturePlot(cd8_pd1, features=c("CD4", "CD8A"), cols=c("grey", "red"))
p2<-FeaturePlot(cd4_pd1, features=c("CD4", "CD8A"), cols=c("grey", "red"))
cowplot::plot_grid(plotlist=list(p1, p2), ncol=1)
ggsave(paste0(figure_dir,"UMAP_CD4.CD8_detailedMarkers_preReannotation.png"), width=7, height=6)

  ## Fix CD4: 1 clear CD8 cluster + 1 with CD4/CD8 (kept as CD4) + 1 cluster annotated cell-wise
  cd4_pd1<- FindClusters(cd4_pd1, resolution = 1.8)
  DimPlot(cd4_pd1, label=TRUE) + VlnPlot(cd4_pd1, features=c("CD4", "CD8A"))
  
    ## clear cluster
    cd4_pd1$detailed_annot_projected_manual<- ifelse(cd4_pd1$seurat_clusters==7, "CD8", cd4_pd1$detailed_annot_projected) 
    
    ## cell wise
    cd4_pd1$cd8_expr<- as.data.frame(cd4_pd1@assays$RNA@counts["CD8A",])
    cd4_pd1$detailed_annot_projected_manual<- ifelse(cd4_pd1$seurat_clusters %in% c(3,6,4,0) & cd4_pd1$cd8_expr > 0, "CD8",
                                                     cd4_pd1$detailed_annot_projected_manual)
  
  DimPlot(cd4_pd1, label=TRUE) + 
    FeaturePlot(cd4_pd1, features=c("CD4", "CD8A"), cols=c("grey", "red"))+
    VlnPlot(cd4_pd1, features=c("CD4", "CD8A"))+
    DimPlot(cd4_pd1, group.by = "detailed_annot_projected_manual")
  ggsave(paste0(figure_dir, "CD4_reannotateToCD8.png") ,width = 10, height=10)
  
  
  ## Fix CD8: 1 clear CD4 cluster, 1 cycling annotated cell wise
  DimPlot(cd8_pd1, label=TRUE) + VlnPlot(cd8_pd1, features=c("CD4", "CD8A"))
  
    ## clear cluster
    cd8_pd1$detailed_annot_projected_manual<- ifelse(cd8_pd1$seurat_clusters==2, "CD4", cd8_pd1$detailed_annot_projected)
    
    ## cell wise
    cd8_pd1$cd4_expr<- as.data.frame(cd8_pd1@assays$RNA@counts["CD4",])
    cd8_pd1$detailed_annot_projected_manual<- ifelse(cd8_pd1$seurat_clusters==3 & cd8_pd1$cd4_expr > 0, "CD4",
                                                     cd8_pd1$detailed_annot_projected_manual)
  
  DimPlot(cd8_pd1, label=TRUE) + 
    FeaturePlot(cd8_pd1, features=c("CD4", "CD8A"), cols=c("grey", "red"))+
    VlnPlot(cd8_pd1, features=c("CD4", "CD8A"))+
    DimPlot(cd8_pd1, group.by = "detailed_annot_projected_manual")
  ggsave(paste0(figure_dir, "CD8_reannotateToCD4.png") ,width = 10, height=10)
  
  ## Add CD4/CD8 to correct object and recluster
  cd8_pd1_reAnnot<- my_mergeRerunClustering(subset(cd8_pd1, detailed_annot_projected_manual=="CD8"),
                                            subset(cd4_pd1, detailed_annot_projected_manual=="CD8"), 
                                            Integration="Harmony", integrateBy = "sample")
  cd4_pd1_reAnnot<- my_mergeRerunClustering(subset(cd8_pd1, detailed_annot_projected_manual=="CD4"),
                                            subset(cd4_pd1, detailed_annot_projected_manual=="CD4"), 
                                            Integration="Harmony", integrateBy = "sample")
  
  ## Confirm finalized annots
  p1<-FeaturePlot(cd8_pd1_reAnnot, features=c("CD4", "CD8A"), cols=c("grey", "red"))
  p2<- FeaturePlot(cd4_pd1_reAnnot, features=c("CD4", "CD8A"), cols=c("grey", "red"))
  cowplot::plot_grid(plotlist=list(p1, p2), ncol=1)
  
  ## Fix cd8 missannotated as cd4- don't cluster prior to reannotation, but do after
  cd4_pd1<- cd4_pd1_reAnnot; cd8_pd1<- cd8_pd1_reAnnot
  cd4_pd1<- FindClusters(cd4_pd1, resolution = 1.5)
  cd4_pd1$cd8_expr<- as.data.frame(cd4_pd1@assays$RNA@counts["CD8A",])
  cd4_pd1$detailed_annot_projected_manual<- ifelse(cd4_pd1$seurat_clusters %in% c(2) & cd4_pd1$cd8_expr > 0, "CD8",
                                                     cd4_pd1$detailed_annot_projected_manual)
 
  
  ## Add CD4/CD8 to correct object and recluster
  cd8_pd1_reAnnot<- my_mergeRerunClustering(subset(cd8_pd1, detailed_annot_projected_manual=="CD8"),
                                            subset(cd4_pd1, detailed_annot_projected_manual=="CD8"), 
                                            Integration="Harmony", integrateBy = "sample")
  cd4_pd1_reAnnot<- subset(cd4_pd1, detailed_annot_projected_manual=="CD4")
  cd4_pd1_reAnnot_meta<-cd4_pd1_reAnnot@meta.data
  cd4_pd1_reAnnot<- RunFullSeurat(cd4_pd1_reAnnot@assays$RNA@counts, RunHarmony = TRUE, samples=cd4_pd1_reAnnot_meta$sample)
  cd4_pd1_reAnnot<- AddMetaData(cd4_pd1_reAnnot, cd4_pd1_reAnnot_meta[,!colnames(cd4_pd1_reAnnot_meta) %in% colnames(cd4_pd1_reAnnot@meta.data)])
  
  ## Confirm finalized annots
  p1<-FeaturePlot(cd8_pd1_reAnnot, features=c("CD4", "CD8A"), cols=c("grey", "red"))
  p2<- FeaturePlot(cd4_pd1_reAnnot, features=c("CD4", "CD8A"), cols=c("grey", "red"))
  cowplot::plot_grid(plotlist=list(p1, p2), ncol=1)
  ggsave(paste0(figure_dir, "UMAP_Tcells_detailedMarkers_postReannotation.png"), width=6, height=6)
  

  
  ## Add back to full tcell
  tcell_pd1$detailed_annot_projected_manual<- ifelse(colnames(tcell_pd1) %in% colnames(cd4_pd1_reAnnot), "CD4",
                                                     ifelse(colnames(tcell_pd1) %in% colnames(cd8_pd1_reAnnot), "CD8", "issue"))
  DimPlot(tcell_pd1, group.by = "detailed_annot_projected_manual")+
    scale_color_manual(values = immune_colors[names(immune_colors) %in% unique(tcell_pd1$detailed_annot_projected_manual)])
  ggsave(paste0(figure_dir, "UMAP_Tcells_detailedAnnot_postReannotation.png"), width=6, height=5)
  

saveRDS(cd4_pd1_reAnnot, file=paste0(analysis_dir, "cd4_seurat.Rds"))
saveRDS(cd8_pd1_reAnnot, file=paste0(analysis_dir, "cd8_seurat.Rds"))
saveRDS(tcell_pd1, file=paste0(analysis_dir, "tcell_seurat.Rds"))
```
