==> PD1_Treat/01a_SS2/Immune/04_ProjectPrograms.Rmd <==
---
title: "Project programs"
author: "Jenna LaBelle"
date: "10/07/22"
output: html_document
---

## Project program annotations from the full immune cohort to tenx sample

```{r}
library(Seurat) 
library(ggalluvial)

base_dir = "/Users/jlabelle/"
script_dir = "Dropbox (Partners HealthCare)/FilbinLab/data_analysis/Shared/Scripts/"
working_dir<- paste0(base_dir, 
                     "Dropbox (Partners HealthCare)/FilbinLab/",
                     "data_analysis/Projects/git_repos/Pedi_Immune/PD1_Treat/01a_SS2/Immune/")

analysis_dir<- paste0(working_dir, "/analysis/project_programs/")
figure_dir<- paste0(working_dir, "/figures/project_programs/")
data_dir<- paste0(working_dir, "analysis/Merge/")
full_cohort_dir<- "../../../ImmuneCellAnalysis/"

if(!dir.exists(figure_dir)){dir.create(figure_dir, recursive = TRUE)}
if(!dir.exists(analysis_dir)){dir.create(analysis_dir, recursive = TRUE)}

preprocessing = paste0(base_dir, script_dir, "single_cell_preprocessing_helper_functions.R")
nmf_helper = paste0(base_dir, script_dir, "NMF_helper_function.R")
source(preprocessing)
source(nmf_helper)
source(paste0(base_dir, script_dir, "seurat_transferAnchors_helper.R"))

```


## Load seurat objects: pd-1 (processed in 01b_Preprocess2QC_ToSeurat.Rmd) and full immune cohort (both myeloid and t cells)

```{r}
## Load merged ss2 (pd1, untreated)
ss2_files<- list.files(data_dir)
ss2_files<- ss2_files[!grepl("tcell", ss2_files)]


ss2_seurat<- lapply(ss2_files, function(x){readRDS(paste0(data_dir, x))})
names(ss2_seurat)<-  gsub("_seurat.Rds", "", ss2_files)
ss2_seurat<- lapply(ss2_seurat, function(x){x$Sequencing<- "SS2"; return(x)})
ss2_seurat$myeloid$Final_Annot<- ss2_seurat$myeloid$NMF_rank6_annotByAdult10X



## Full cohort seurat objects
full_myeloid_seurat<- readRDS(paste0(full_cohort_dir,
                                     "01_Preprocessing/02_Counts_Preprocess/analysis/detailed_annotation/",
                                     "pedOnly_nomuv63/",
                                     "myeloid_seurat_noDC.B_harmony.Rds"))
full_myeloid_seurat$Final_Annot<-full_myeloid_seurat$NMF_rank6_annotByAdult10X
full_cd4_seurat<- readRDS(paste0(full_cohort_dir,
                                   "02a_Tcells/02_Ped.Adult/03_Visualize.Reannotate_programs/analysis/Reannotate_programs/",
                                   "rank8_sharedByunion/",
                                   "ped_seurat_cd4_harmony.Rds"))
full_cd8_seurat<- readRDS(paste0(full_cohort_dir,
                                   "02a_Tcells/02_Ped.Adult/03_Visualize.Reannotate_programs/analysis/Reannotate_programs/",
                                   "rank8_sharedByunion/",
                                   "ped_seurat_cd8_harmony.Rds"))

## Colors 
tcell_colors<- readRDS(paste0(full_cohort_dir, "/plot_colors/Tcell.programs_rank8sharedByunion_reAnnot.Rds"))
myeloid_colors<- readRDS(paste0(full_cohort_dir, "/plot_colors/Myeloid_program_colors.Rds"))
myeloid_colors<- myeloid_colors$rank6
immune_colors<- c(tcell_colors, myeloid_colors)
colors_treatment<- c(PD1="orange", Untreated="purple")

## Full cohort marker genes
tcell_markers<- readRDS(paste0(full_cohort_dir, 
                             "02a_Tcells/02_Ped.Adult/01_compare_NMF/analysis/newMarkers_deNovoPed.AdultTcell/", 
                             "new_markers_NMFrank8_bySeurat_sharedByunion_reAnnot.Rds"))
myeloid_markers<- read.csv(paste0(full_cohort_dir,
                               "02b_Myeloid/01a_NMF/analysis/de.novoNMF/pedOnly_nomuv63/", 
                               "top_DE_genes_10Xannotated_NMF_rank6.csv"))

  ## Top markers per program
  top_tcell_markers<- tcell_markers %>% group_by(cluster) %>% top_n(n=5, wt=avg_log2FC) %>% as.data.frame()
  top_myeloid_markers<- myeloid_markers %>% group_by(cluster) %>% top_n(n=5, wt=avg_log2FC) %>% as.data.frame()

```

## To assess accuracy of projection: (not actually used downstream)
Remove GBM samples from full cohort and project
Does this projection still result in "good" (i.e., better than scoring and assigning) program assignment?
Results: assignment accuracy worse than projection with full cohort (obviously) but still much improved compared to scoring method. Projection with full cohort used downstream.
```{r}
## Remove gbm samples from full cohort seurat objects to be used as reference
all_samples<- unique(c(full_myeloid_seurat$sample, full_cd4_seurat$sample, full_cd8_seurat$sample))
gbm_samples<- unique(c(ss2_seurat$cd4$sample))
non_gbm_samples<- all_samples[!all_samples %in% gbm_samples]

myeloid_ref<- subset(full_myeloid_seurat, sample %in% non_gbm_samples)
cd4_ref<- subset(full_cd4_seurat, sample %in% non_gbm_samples)
cd8_ref<- subset(full_cd8_seurat, sample %in% non_gbm_samples)

## Project
ss2_seurat$cd4<- my_project_seurat(cd4_ref,ss2_seurat$cd4, reference_annotation = "NMF_rank8_annot_mergedByunion_reAnnot")
ss2_seurat$cd8<- my_project_seurat(cd8_ref,ss2_seurat$cd8, reference_annotation = "NMF_rank8_annot_mergedByunion_reAnnot")
ss2_seurat$myeloid<- my_project_seurat(myeloid_ref,ss2_seurat$myeloid, reference_annotation = "NMF_rank6_annotByAdult10X")

## Compare program assignment to actual program assignment
all_plots<- lapply(names(ss2_seurat), function(x){
  print(x)
  tmp<- ss2_seurat[[x]]@meta.data
  bd<- as.data.frame(table(tmp$Final_Annot, tmp$Program_projected))
  colnames(bd)<-c("OriginalAnnotation", "ProjectedAnnotation", "NumberOfCells")
  my_sankeyPlot(bd, immune_colors) + ggtitle(x)
})
cowplot::plot_grid(plotlist=all_plots, ncol=1)
ggsave(paste0(figure_dir, "SS2_Sankey_ChangeInAnnotation_Original.v.Projected_noGBMInRef.png"), width=6, height=12)
```


## Project programs from full cohort to gbm (ss2 and 10x, with treated/untreated together)
```{r}
## SS2
ss2_seurat$cd4<- my_project_seurat(full_cd4_seurat,ss2_seurat$cd4, reference_annotation = "NMF_rank8_annot_mergedByunion_reAnnot")
ss2_seurat$cd8<- my_project_seurat(full_cd8_seurat,ss2_seurat$cd8, reference_annotation = "NMF_rank8_annot_mergedByunion_reAnnot")
ss2_seurat$myeloid<- my_project_seurat(full_myeloid_seurat,ss2_seurat$myeloid, reference_annotation = "NMF_rank6_annotByAdult10X")


for(i in names(ss2_seurat)){
  saveRDS(ss2_seurat[[i]], file=paste0(data_dir, i, "_seurat.Rds"))
}

```


## For samples from full cohort, plot change in annotations (i.e., accuracy)
```{r}
all_plots<- lapply(names(ss2_seurat), function(x){
  tmp<- ss2_seurat[[x]]@meta.data
  bd<- as.data.frame(table(tmp$Final_Annot, tmp$Program_projected))
  colnames(bd)<-c("OriginalAnnotation", "ProjectedAnnotation", "NumberOfCells")
  my_sankeyPlot(bd, immune_colors) + ggtitle(x)
})
cowplot::plot_grid(plotlist=all_plots, ncol=1)
ggsave(paste0(figure_dir, "SS2_Sankey_ChangeInAnnotation_Original.v.Projected.png"), width=6, height=12)
```


## UMAPs of projected programs
```{r}

all_plots<- lapply(names(ss2_seurat), function(x){
  tmp<- ss2_seurat[[x]]
  DimPlot(tmp, group.by="Program_projected", pt.size = 1, label=FALSE, label.size = 6)+ 
    scale_color_manual(values=immune_colors[names(immune_colors) %in% tmp$Program_projected])+
    ggtitle(x)
})
cowplot::plot_grid(plotlist=all_plots, ncol=3)
ggsave(paste0(figure_dir, "SS2_UMAP_PredictedProgram.png"), width=18, height=5)

```

## Barchart of program proportions
```{r}

all_plots<- lapply(names(ss2_seurat), function(x){
  tmp<- ss2_seurat[[x]]
  my_barchart_programProp(tmp, coloring_variable = "Program_projected", colors=immune_colors)+ggtitle(x)
})
cowplot::plot_grid(plotlist=all_plots, ncol=3)
ggsave(paste0(figure_dir, "SS2_Barchart_PredictedProgram.png"), width=18, height=5)

```


## NOT RUN
## Pseudobulked heatmap of expression of marker genes
```{r}
marker_list<- list(cd4=top_tcell_markers[!top_tcell_markers$cluster %in% c("Predysfunctional", "Cytotoxic.NKreceptor"),], 
                   cd8=top_tcell_markers[!top_tcell_markers$cluster %in% c("Treg", "Inhibitory"),], 
                   myeloid=top_myeloid_markers)

## Merge (no clustering needed) across sequencing types
all_plots<- lapply(gsub("tenx_", "", names(ss2_seurat)), function(x){
  tmp<- merge(ss2_seurat[[paste0("tenx_", x)]], ss2_seurat[[paste0("ss2", x)]])
  markers_use<-marker_list[[x]]
  markers_use<- markers_use[markers_use$cluster %in% unique(tmp$Program_projected),]
  
  ## Create pseudobulking variable
  tmp$pb<- paste0(tmp$sample, "_", gsub("\\.", "\n", gsub("_", "\n", tmp$Program_projected)))
  pb<- pseudobulk_byVariable(tmp, tmp@meta.data, "pb")
  
  
  ## center for each sample separately
  center_bySample<- lapply(unique(tmp$sample), function(sample){
    pb_tmp<- pb[,grepl(sample, colnames(pb))]
    if(class(pb_tmp)=="numeric"){
      pb_tmp<- as.data.frame(pb_tmp)
      colnames(pb_tmp)<- colnames(pb)[grepl(sample, colnames(pb))]
    }
    pb_list_tmp<- NormCenter(pb_tmp)
    pb_center_tmp<- pb_list_tmp$center_data
    return(pb_center_tmp)
  })
  center_bySample<- do.call("cbind", center_bySample)

  
  ## Plot
  myHeatmap(center_bySample, GOI=markers_use$gene, min.value = -3, max.value = 3, 
                orderFactors =  gsub("\\.", "\n",gsub("_", "\n", unique(markers_use$cluster))),
            orderSubtypes =c("BT1478", "BT1857","BT1745", "E167", "BT1910", "BT1935", "GEX") ) +
    scale_fill_gradient2(high="red", low="blue", mid="white", name="Centered log2(TPM)")+
    ggtitle(x)
})
cowplot::plot_grid(plotlist=all_plots, ncol=3)
ggsave(paste0(figure_dir, "Heatmap_ExpectedMarkerGenes_All.png"), width=35, height=9)
```

## DE between programs by treatment
Separately for SS2/10x, then look for overlaps
```{r}
## SS2
ss2_de<-lapply(ss2_seurat, function(seurat){
  ## Identify programs to test- >3 cells in program in both treated/untreated
  programs_in_both<- table(seurat$Treatment, seurat$Program_projected)>3
  programs_in_both<- colnames(programs_in_both)[colSums(programs_in_both)==2]
  message(paste0("Programs testing: ", paste(programs_in_both, collapse = ", ")))
  
  ## Cycle through each program shared by treated/untreated (>3 cells in program)
  programs_de<- lapply(programs_in_both, function(program){
    print(program)
    seurat_subset<- subset(seurat, Program_projected==program)
    seurat_subset<- SetIdent(seurat_subset, value=seurat_subset$Treatment)
    markers<- FindMarkers(seurat_subset, ident.1 = "Untreated", ident.2="PD1")
    markers$program<- program
    markers$gene<- rownames(markers)
    return(markers)
  })
  programs_de<- do.call("rbind", programs_de)
  
  return(programs_de)
})

  ## Remove any cell types with no degs
  ss2_de<- ss2_de[lapply(ss2_de, class)=="data.frame"]
  
  ## filter
  ss2_de<- lapply(ss2_de, function(x){
    x[x$p_val_adj<0.1,]
  })
  
  saveRDS(ss2_de, file=paste0(analysis_dir, "DE_Untreated.v.PD1_byProgram_SS2.Rds"))
  
  
## tenx
tenx_de<-lapply(ss2_seurat, function(seurat){
  ## Identify programs to test- >3 cells in program in both treated/untreated
  programs_in_both<- table(seurat$Treatment, seurat$Program_projected)>3
  programs_in_both<- colnames(programs_in_both)[colSums(programs_in_both)==2]
  message(paste0("Programs testing: ", paste(programs_in_both, collapse = ", ")))
  
  ## Cycle through each program shared by treated/untreated (>3 cells in program)
  programs_de<- lapply(programs_in_both, function(program){
    print(program)
    seurat_subset<- subset(seurat, Program_projected==program)
    seurat_subset<- SetIdent(seurat_subset, value=seurat_subset$Treatment)
    markers<- FindMarkers(seurat_subset, ident.1 = "Untreated", ident.2="PD1")
    markers$program<- program
    markers$gene<- rownames(markers)
    return(markers)
  })
  programs_de<- do.call("rbind", programs_de)
  
  return(programs_de)
})

  ## Remove any cell types with no degs
  tenx_de<- tenx_de[lapply(tenx_de, class)=="data.frame"]
  
  ## filter
  tenx_de<- lapply(tenx_de, function(x){
    x[x$p_val_adj<0.1,]
  })
  
  saveRDS(tenx_de, file=paste0(analysis_dir, "DE_Untreated.v.PD1_byProgram_TenX.Rds"))

## Merge results by cell type
all_degs<- lapply(gsub("ss2", "", names(ss2_de)), function(x){
  ss2<- ss2_de[[paste0("ss2", x)]]
  tenx<- tenx_de[[paste0("tenx_", x)]]
  
  ## Merge DE results
  ss2$Seq<- "SS2"; tenx$Seq<- "TenX"
  de_both<- rbind(ss2, tenx)
  return(de_both)
}); names(all_degs)<-gsub("ss2", "", names(ss2_de))
  
## Venn diagram of overlaps
all_plots<-lapply(gsub("ss2", "", names(ss2_de)), function(x){
  de_both<-all_degs[[x]]
  cowplot::plot_grid(plotlist=lapply(unique(de_both$program), function(program){
    tmp<- de_both[de_both$program==program,]
    ggVennDiagram::ggVennDiagram(list(ss2=tmp$gene[tmp$Seq=="SS2"],
                                      tenx=tmp$gene[tmp$Seq=="TenX"]))+
      scale_fill_gradient(low="white", high="red")+
      scale_color_manual(values=rep("black",3))+
      ggtitle(program)
  }))
  ggsave(paste0(figure_dir, "Venn_DEGoverlap_SS2.v.TenX_", x, ".png"), width=9, height=6)
})
  
## Get overlapping genes
degs_shared<- lapply(gsub("ss2", "", names(ss2_de)), function(x){
  de_both<- all_degs[[x]]
  
  ## Get genes that are DE in both
  genes_shared<- as.data.frame(t(table(de_both$program, de_both$gene)))
  genes_shared<- genes_shared[genes_shared$Freq==2,]
  de_shared<- de_both[paste0(de_both$gene, de_both$program) %in% paste0(genes_shared$Var1, genes_shared$Var2),]
  
  ## Subset to genes with same logFC
  de_shared<- de_shared %>% group_by(program, gene) %>% filter(length(unique(sign(avg_log2FC))) == 1 ) %>% as.data.frame()
  de_shared<- de_shared[order(de_shared$program, de_shared$gene),]
  
  ## Reverse logFC sign- so that positive == up in PD1
  de_shared$avg_log2FC<- de_shared$avg_log2FC * -1
  de_shared$cell_type<- x
  return(de_shared)
})
degs_shared<- do.call("rbind", degs_shared)


## List of GOI of shared DEGs (based on lit review)
GOI<- c("GADD45B", "IL6ST", "TOB1", "HAMP", "LYVE1", "NINJ1", "RHOB", "SLC2A5", "STAB1", "IL32")
names(GOI)<- plyr::mapvalues(GOI, degs_shared$gene, degs_shared$program, warn_missing = FALSE)

## Plot expression for ss2/tenx for all goi
all_plots<- lapply(gsub("tenx_", "", names(ss2_seurat)), function(x){
  tmp<- all_seurat[[x]]
  goi_use<- GOI[names(GOI) %in% tmp$Program_projected]
  
  ## Create pseudobulking variable
  tmp$pb<- paste0(tmp$sample, "_", gsub("\\.", "\n", gsub("_", "\n", tmp$Program_projected)))
  pb<- pseudobulk_byVariable(tmp, tmp@meta.data, "pb")
  
  pb_center<- NormCenter(pb)$center_data
  
  ## Plot
  myHeatmap(pb_center, GOI=goi_use, min.value = -3, max.value = 3, facetWrap = TRUE,
            orderSubtypes =sample_order) +
    scale_fill_gradient2(high="red", low="blue", mid="white", name="Centered log2(TPM)")+
    ggtitle(x)
})
cowplot::plot_grid(plotlist=all_plots, ncol=1)
ggsave(paste0(figure_dir, "Heatmap_DEGsSharedBySS2.TenX_PD1.v.Untreated_All.png"), width=12, height=12)
```



## For visualization purposes: map 10x/treated onto full seurat object 
For full immune cohort T cells: can use either ped/adult or just ped. Visualization is nicer with ped/adult, but should probably use ped alone for actual projections (keep it more of a direct comparison, plus doesn't make sense for myeloid, so want to treat tcell/myeloid the same)
```{r}
## Load in tcell seurat objects- pediatric and adult
full_cd4_seurat<- readRDS(paste0(full_cohort_dir,
                                   "02a_Tcells/02_Ped.Adult/03_Visualize.Reannotate_programs/analysis/Reannotate_programs/",
                                   "rank8_sharedByunion/",
                                   "ped.adult_seurat_cd4_harmony.Rds"))
full_cd8_seurat<- readRDS(paste0(full_cohort_dir,
                                   "02a_Tcells/02_Ped.Adult/03_Visualize.Reannotate_programs/analysis/Reannotate_programs/",
                                   "rank8_sharedByunion/",
                                   "ped.adult_seurat_cd8_harmony.Rds"))

## Named list of objects to help with iterating over
ref_list<- list(cd4=full_cd4_seurat, cd8=full_cd8_seurat, myeloid=full_myeloid_seurat)
ss2pd1_list<- list(cd4=subset(ss2_seurat$cd4, Treatment=="PD1"),
                    cd8=subset(ss2_seurat$cd8, Treatment=="PD1"),
                    myeloid=subset(ss2_seurat$myeloid, Treatment=="PD1"))
tenx_list<- ss2_seurat; names(tenx_list)<-gsub("tenx_", "", names(tenx_list))
color_list<- list(cd4=tcell_colors, cd8=tcell_colors, myeloid=myeloid_colors)

## SS2 PD1
ss2pd1_list<- lapply(names(ss2pd1_list), function(x){
  print(x)
  my_mapSeurat(ref_list[[x]], ss2pd1_list[[x]], metadata_column_to_project = "Final_Annot" )
})
names(ss2pd1_list)<- names(ref_list)


  ## Plot by program (reference + query)
  all_plots<- lapply(names(ss2pd1_list), function(x){
    ref<- ref_list[[x]]; query<- ss2pd1_list[[x]]
    
    p1<-DimPlot(ref, group.by = "Final_Annot") + 
      scale_color_manual(values=color_list[[x]])+
      ggtitle(paste0(x, " :Full cohort"))
    
    p2<- DimPlot(query, reduction="ref.umap", group.by = "Final_Annot") + 
      scale_color_manual(values=color_list[[x]])+
      ggtitle(paste0(x, " : PD1 SS2"))
    p1+p2
  })
  cowplot::plot_grid(plotlist = all_plots, ncol=1)
  ggsave(paste0(figure_dir, "UMAP_ss2_mapOntoFull_byProgram_All.png"), width=14, height=12)
  
  ## breakdown by program
  all_plots<- lapply(names(ss2pd1_list), function(x){
    my_barchart_programProp(ss2pd1_list[[x]], coloring_variable="predicted.Final_Annot", colors=color_list[[x]])+ggtitle(x)
  })
  cowplot::plot_grid(plotlist = all_plots, ncol=3)
  ggsave(paste0(figure_dir, "Barchart_ss2_mapOntoFull_byProgram_All.png"), width=12, height=4)

  
## 10x- pd1 and nontreated
tenx_list<- lapply(names(tenx_list), function(x){
  print(x)
  my_mapSeurat(ref_list[[x]], tenx_list[[x]], metadata_column_to_project = "Final_Annot" )
})  
names(tenx_list)<- gsub("tenx_", "", names(ss2_seurat))
    

  ## Plot by program (reference + query)
  all_plots<- lapply(names(tenx_list), function(x){
    ref<- ref_list[[x]]; query<- tenx_list[[x]]
    
    p1<-DimPlot(ref, group.by = "Final_Annot") + 
      scale_color_manual(values=color_list[[x]])+
      ggtitle(paste0(x, ": Full cohort"))
    
    p2<- DimPlot(query, reduction="ref.umap", group.by = "Final_Annot") + 
      scale_color_manual(values=color_list[[x]])+
      ggtitle(paste0(x, ": 10X"))
    p1+p2
  })
  cowplot::plot_grid(plotlist = all_plots, ncol=1)
  ggsave(paste0(figure_dir, "UMAP_tenX_mapOntoFull_byProgram_All.png"), width=14, height=12)

  
  ## Plot by treatment
  all_plots<- lapply(names(tenx_list), function(x){
    query<- tenx_list[[x]]
    DimPlot(query, group.by = "Treatment", reduction="ref.umap")+scale_color_manual(values=colors_treatment)+
      ggtitle(paste0(x, " : 10X2"))
  })
  cowplot::plot_grid(plotlist = all_plots, ncol=1)
  ggsave(paste0(figure_dir, "UMAP_tenX_mapOntoFull_byTreatment_All.png"),width=6, height=15)
  
    ## breakdown by program
  all_plots<- lapply(names(tenx_list), function(x){
    my_barchart_programProp(tenx_list[[x]], coloring_variable="predicted.Final_Annot", colors=color_list[[x]])+ggtitle(x)
  })
  cowplot::plot_grid(plotlist = all_plots, ncol=3)
  ggsave(paste0(figure_dir, "Barchart_tenx_mapOntoFull_byProgram_All.png"), width=12, height=4)
  

## Breakdown merged across all- including gbm from full cohort
gbm_samples<- c("BT1857", "BT1478", "BT1745")

  ## Merge seurat objects- no clustering, not needed
  all_seurat_list<- lapply(names(ref_list), function(x){
    ref<- ref_list[[x]]; ref<- subset(ref, sample %in% gbm_samples)
    ref$Treatment<- "Untreated"; ref$Sequencing<- "SS2"; ref$predicted.Final_Annot<- ref$Final_Annot
    ss2<-ss2pd1_list[[x]]
    tenx<- tenx_list[[x]]
    seurat<- merge(ref, c(tenx, ss2))
    return(seurat)
  })
  names(all_seurat_list)<- names(ref_list)
  
  ## Bar chart
  sample_order<- c("BT1478", "BT1857", "BT1745", "E167CD35resub", "BT1910", "BT1935", "GEX")
  all_plots<- lapply(names(all_seurat_list), function(x){
    seurat<- all_seurat_list[[x]]
    seurat$sample<- factor(seurat$sample, levels=sample_order[sample_order %in% unique(seurat$sample)])
    my_barchart_programProp(seurat, coloring_variable="predicted.Final_Annot", colors=color_list[[x]])+
      theme(axis.text.x = element_text(angle=45, hjust=1))+xlab("")
    
  })
  cowplot::plot_grid(plotlist=all_plots, ncol=3)
  ggsave(paste0(figure_dir, "Barchart_tenx.ss2_mapOntoFull_byProgram_All.png"), width=12, height=4)
  
  ## Bar chart averaged over treatment
    all_plots<- lapply(names(all_seurat_list), function(x){
    seurat<- all_seurat_list[[x]]
    seurat$sample<- factor(seurat$sample, levels=sample_order[sample_order %in% unique(seurat$sample)])
    my_barchart_programProp(seurat,grouping_variable = "Treatment", coloring_variable="predicted.Final_Annot", colors=color_list[[x]])+
      theme(axis.text.x = element_text(angle=45, hjust=1))+xlab("")
    
  })
  cowplot::plot_grid(plotlist=all_plots, ncol=3)
  ggsave(paste0(figure_dir, "Barchart_tenx.ss2_mapOntoFull_byProgram_byTreatment_All.png"), width=12, height=4)

```
